since multithreading turned out to be bad with cpu heavy applications due to the python GIL, i changed multithreading to multiprocessing which is better for cpu heavy parallelisation; the rebuilding was moved to an external file to avoid pickling problems
